fix bug showing network streamlit
